requirements.txt
requests==2.23.0
python-telegram-bot==12.4.2

